+++ src/main/java/com/fiseq/truckcompany/entities/UserProfile.java
@@ -28,7 +28,7 @@ public class UserProfile implements Serializable {
     @OneToMany(mappedBy = "owner", cascade = CascadeType.ALL)
     private List<Job> jobs;
 
-    private double totalMoney;
+    private Integer totalMoney;
 
 }
